refactor google doc test scripts
